muberra commits, fixes and reformatting

+++ backend/models/upcoming_payment.go
@@ -1,12 +1,12 @@
 package models
 
 type UpcomingPayment struct {
-	Name          string `json:"name"`
-	Price         int    `json:"price"`
-	RenewalPeriod string `json:"renewalPeriod"`
-	Repetition    int    `json:"repetition"`
-	Date          string `json:"date"`
-	Reminder      string `json:"reminder"`
-	Category      string `json:"category"`
-	Description   string `json:"description"`
+	Name          string  `json:"name"`
+	Price         float64 `json:"price"`
+	RenewalPeriod string  `json:"renewalPeriod"`
+	Repetition    int     `json:"repetition"`
+	Date          string  `json:"date"`
+	Reminder      string  `json:"reminder"`
+	Category      string  `json:"category"`
+	Description   string  `json:"description"`
 }
